Fix cards reducer
